Skip non Algolia entities in commands
When invoking the clear or reindex command for all known entities
the command will fail if at least one entity is not Algolia mapped.
This is incosistent regarding the settings command and makes it
difficult in automation processes.

With this change the commands will no longer trigger an error
but skip non Algolia entities. An informational output will be shown.